_back02keystone: uuiding() plugin
