+++ apps/_back02keystone/custom-plugins/_common.js
@@ -1,0 +1,11 @@
+const composeHook = (originalHook, newHook) => async params => {
+    let { resolvedData } = params
+    if (originalHook) {
+        resolvedData = await originalHook(params)
+    }
+    return newHook({ ...params, resolvedData })
+}
+
+module.exports = {
+    composeHook,
+}
