+++ apps/_back02keystone/custom-plugins/index.js
@@ -1,7 +1,9 @@
 const { versioning } = require('./versioning')
 const { historical } = require('./historical')
+const { uuiding } = require('./uuiding')
 
 module.exports = {
     versioning,
     historical,
+    uuiding,
 }
