+++ apps/_back02keystone/custom-plugins/uuiding.js
@@ -1,0 +1,18 @@
+const { v4: uuid } = require('uuid')
+const { Uuid } = require('@keystonejs/fields')
+
+const uuiding = () => ({ fields = {}, hooks = {}, ...rest }) => {
+    const idOptions = {
+        type: Uuid,
+        defaultValue: () => uuid(),
+        isRequired: true,
+    }
+
+    fields['id'] = idOptions
+
+    return { fields, hooks, ...rest }
+}
+
+module.exports = {
+    uuiding,
+}
